revert to old implementation and fixes
