--- cypress/fixtures/user_uuid_valC353540.csv
@@ -1,3 +1,0 @@
-10cb9367-c095-4872-9add-8aecdf339dd4
-888a321d-676e-42fc-b588-a677d16a76ec
-5cc5bd09-d90e-4484-8058-74c237165877
